Extract `substituteAtApply` into separate file
