+++ packages/tailwindcss/src/index.ts
@@ -1,11 +1,11 @@
 import { version } from '../package.json'
-import { WalkAction, comment, decl, rule, toCss, walk, type AstNode, type Rule } from './ast'
+import { substituteAtApply } from './apply'
+import { WalkAction, comment, decl, rule, toCss, walk, type Rule } from './ast'
 import { compileCandidates } from './compile'
 import * as CSS from './css-parser'
 import { buildDesignSystem, type DesignSystem } from './design-system'
 import { buildPluginApi, type PluginAPI } from './plugin-api'
 import { Theme } from './theme'
-import { escape } from './utils/escape'
 import { segment } from './utils/segment'
 
 const IS_VALID_UTILITY_NAME = /^[a-z][a-zA-Z0-9/%._-]*$/
@@ -383,72 +383,6 @@ export async function compile(
       return compiledCss
     },
   }
-}
-
-function substituteAtApply(ast: AstNode[], designSystem: DesignSystem) {
-  walk(ast, (node, { replaceWith }) => {
-    if (node.kind !== 'rule') return
-    if (!(node.selector[0] === '@' && node.selector.startsWith('@apply '))) return
-
-    let candidates = node.selector
-      .slice(7 /* Ignore `@apply ` when parsing the selector */)
-      .trim()
-      .split(/\s+/g)
-
-    // Replace the `@apply` rule with the actual utility classes
-    {
-      // Parse the candidates to an AST that we can replace the `@apply` rule
-      // with.
-      let candidateAst = compileCandidates(candidates, designSystem, {
-        onInvalidCandidate: (candidate) => {
-          throw new Error(`Cannot apply unknown utility class: ${candidate}`)
-        },
-      }).astNodes
-
-      // Collect the nodes to insert in place of the `@apply` rule. When a rule
-      // was used, we want to insert its children instead of the rule because we
-      // don't want the wrapping selector.
-      let newNodes: AstNode[] = []
-      for (let candidateNode of candidateAst) {
-        if (candidateNode.kind === 'rule' && candidateNode.selector[0] !== '@') {
-          for (let child of candidateNode.nodes) {
-            newNodes.push(child)
-          }
-        } else {
-          newNodes.push(candidateNode)
-        }
-      }
-
-      // Verify that we don't have any circular dependencies by verifying that
-      // the current node does not appear in the new nodes.
-      walk(newNodes, (child) => {
-        if (child !== node) return
-
-        // At this point we already know that we have a circular dependency.
-        //
-        // Figure out which candidate caused the circular dependency. This will
-        // help to create a useful error message for the end user.
-        for (let candidate of candidates) {
-          let selector = `.${escape(candidate)}`
-
-          for (let rule of candidateAst) {
-            if (rule.kind !== 'rule') continue
-            if (rule.selector !== selector) continue
-
-            walk(rule.nodes, (child) => {
-              if (child !== node) return
-
-              throw new Error(
-                `You cannot \`@apply\` the \`${candidate}\` utility here because it creates a circular dependency.`,
-              )
-            })
-          }
-        }
-      })
-
-      replaceWith(newNodes)
-    }
-  })
 }
 
 export function __unstable__loadDesignSystem(css: string) {
